+++ groovyConsole/src/main/groovy/GroovyConsole.groovy
@@ -294,13 +294,13 @@ class GroovyConsole {
         def writeButton = swing.button(
                 text : showLabels? 'send to node' : null ,
                 toolTipText : 'save script into node',
-                icon : menuUtils.getMenuItemIcon('IconAction.emoji-2B07'),
+                icon : menuUtils.getMenuItemIcon('IconAction.groovyConsole/saveInNode'),
                 actionPerformed : {e -> write(console)},
              )
         def writeAsButton = swing.button(
                 text : showLabels? 'config + send to node' : null ,
                 toolTipText : 'select options and save script into node',
-                icon : menuUtils.getMenuItemIcon('IconAction.emoji-2198'),
+                icon : menuUtils.getMenuItemIcon('IconAction.groovyConsole/saveInNodeAs'),
                 actionPerformed : {e -> writeTo(console)},
              )
         def refreshCompilerConfiguration = swing.button(
@@ -318,7 +318,7 @@ class GroovyConsole {
         def reLoadButton = swing.button(
                 text : showLabels? 'reload from file' : null ,
                 toolTipText : 'reload script from file',
-                icon : menuUtils.getMenuItemIcon('IconAction.emoji-1F4C1'),
+                icon : menuUtils.getMenuItemIcon('IconAction.groovyConsole/reloadFile'),
                 actionPerformed : {e -> reLoadScriptFile(console)},
              )